All post migrations run in background
- implement parallel background execution of post migration procedures
- implement background job to keep connection opened
- add wait function to avoid background jobs in testing
- move migration functions from database.go to migration.go

On-behalf-of: SAP Michal Krzyz <michal.krzyz@sap.com>
